+++ componentsV2/src/androidTest/java/uk/gov/android/ui/componentsV2/text/GdsAnnotatedStringTest.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.test.core.app.ApplicationProvider
 import org.junit.Rule
 import org.junit.Test
-import uk.gov.android.ui.components.R
+import uk.gov.android.ui.componentsV2.R
 
 class GdsAnnotatedStringTest {
     private val context: Context = ApplicationProvider.getApplicationContext()
