+++ componentsV2/src/main/java/uk/gov/android/ui/componentsV2/GdsContentTile.kt
@@ -26,7 +26,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import uk.gov.android.ui.components.R
 import uk.gov.android.ui.componentsV2.button.ButtonParameters
 import uk.gov.android.ui.componentsV2.button.ButtonType
 import uk.gov.android.ui.componentsV2.button.GdsButton
