+++ componentsV2/src/main/java/uk/gov/android/ui/componentsV2/images/GdsVectorImage.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import uk.gov.android.ui.components.R
+import uk.gov.android.ui.componentsV2.R
 import uk.gov.android.ui.theme.GdsTheme
 
 @Composable
